Rename packaging._compat -> packaging._typing
