+++ src/packaging/_typing.py
